requirements.txt
anthropic
openai
elevenlabs
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
moviepy
pillow
python-dotenv
requests
